refactor: extract cors sanitisation into dedicated service
